pyproject.toml
[project]
name = "carouselmaker"
version = "0.1.0"
description = "Nano Banana Carousel Maker — AI-powered carousel generation for LinkedIn & Instagram"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Web framework
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    # Telegram bot
    "aiogram>=3.17,<4",
    # Database
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "psycopg2-binary>=2.9,<3",
    "alembic>=1.14,<2",
    # Task queue
    "celery[redis]>=5.4,<6",
    # AI providers
    "anthropic>=0.42,<1",
    "google-genai>=1.0,<2",
    # Image rendering
    "Pillow>=11,<12",
    # S3 storage
    "boto3>=1.35,<2",
    # Config
    "pydantic-settings>=2.7,<3",
    # Redis (for bot FSM storage + rate limiting)
    "redis>=5.2,<6",
    # HTTP client
    "httpx>=0.28,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "pytest-cov>=6,<7",
    "ruff>=0.9,<1",
    "mypy>=1.14,<2",
    "factory-boy>=3.3,<4",
    "aiosqlite>=0.20,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.13"
plugins = ["pydantic.mypy"]
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "boto3.*",
    "botocore.*",
    "google.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
